+++ pyvista/core/datasetattributes.py
@@ -4,7 +4,7 @@ from typing import Any, Dict, Iterator, List, Optional, Sequence, Tuple, Union
 
 import numpy as np
 
-from pyvista import _vtk
+import pyvista._vtk_core as _vtk
 import pyvista.utilities.helpers as helpers
 from pyvista.utilities.helpers import FieldAssociation
 from pyvista.utilities.misc import copy_vtk_array
