+++ pyvista/core/filters/structured_grid.py
@@ -2,9 +2,10 @@
 import numpy as np
 
 import pyvista
-from pyvista import _vtk, abstract_class
+import pyvista._vtk_core as _vtk
 from pyvista.core.filters import _get_output
 from pyvista.core.filters.data_set import DataSetFilters
+from pyvista.utilities import abstract_class
 
 
 @abstract_class
