+++ pyvista/core/filters/uniform_grid.py
@@ -4,10 +4,11 @@ import collections.abc
 import numpy as np
 
 import pyvista
-from pyvista import _vtk, abstract_class
+import pyvista._vtk_core as _vtk
 from pyvista.core.filters import _get_output, _update_alg
 from pyvista.core.filters.data_set import DataSetFilters
 from pyvista.errors import AmbiguousDataError, MissingDataError
+from pyvista.utilities import abstract_class
 
 
 @abstract_class
