+++ pyvista/utilities/helpers.py
@@ -31,6 +31,37 @@ class FieldAssociation(enum.Enum):
     ROW = _vtk.vtkDataObject.FIELD_ASSOCIATION_ROWS
 
 
+class AnnotatedIntEnum(int, enum.Enum):
+    """Annotated enum type."""
+
+    def __new__(cls, value, annotation):
+        """Initialize."""
+        obj = int.__new__(cls, value)
+        obj._value_ = value
+        obj.annotation = annotation
+        return obj
+
+    @classmethod
+    def from_str(cls, input_str):
+        """Create from string."""
+        for value in cls:
+            if value.annotation.lower() == input_str.lower():
+                return value
+        raise ValueError(f"{cls.__name__} has no value matching {input_str}")
+
+    @classmethod
+    def from_any(cls, value):
+        """Create from string, int, etc."""
+        if isinstance(value, cls):
+            return value
+        elif isinstance(value, int):
+            return cls(value)
+        elif isinstance(value, str):
+            return cls.from_str(value)
+        else:
+            raise ValueError(f"{cls.__name__} has no value matching {value}")
+
+
 def get_vtk_type(typ):
     """Look up the VTK type for a given numpy data type.
